xen/arm/irq: allow assignment/releasing of eSPI interrupts
The current checks don't allow us to assign or release interrupts
with INTID greater than 1024. This patch adds an additional condition
to check whether the IRQ number is in the eSPI range and allows it to
be assigned to Xen and domains if it is.

Signed-off-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>

modify check in release_guest_irq

Signed-off-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>