xen/arm: gicv3: modify ICH_LR_PHYSICAL_MASK to allow eSPI processing
To properly deactivate guest interrupts and allow them to be retriggered
after the initial trigger, the LR needs to be updated.
The current implementation ignores interrupts outside the range specified
by the mask 0x3FF, which only covers IRQ numbers up to 1023. To enable
processing of eSPI interrupts, this patch updates the mask to 0x13FF.

Signed-off-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>